requirements.txt
Django==5.2.1
python-dotenv
djangorestframework
django-cors-headers
django-filter
psycopg2-binary
gunicorn
whitenoise
dj-database-url
drf-spectacular
Pillow
python-slugify
celery
redis
django-celery-results
django-celery-beat
django-imagekitio-storage==0.0.1
imagekitio==4.2.0
deep-translator